Sync update

+++ src/alpamayo_r1/models/action_in_proj.py
@@ -163,4 +163,4 @@ class PerWaypointActionInProjV2(torch.nn.Module):
         timestep_feats = self.timestep_fourier_encoder(timesteps[..., -1])
         timestep_feats = timestep_feats.repeat(1, T, 1)
         x = torch.cat((action_feats, timestep_feats), dim=-1)
-        return self.norm(self.encoder(x.flatten(0, 1)).reshape(B, T, -1))
+        return self.norm(self.encoder(x.flatten(0, 1)).reshape(B, T, -1)).to(dtype=x.dtype)
